azure-validation-toolkit/requirements.txt
azure-cli
azure-identity
azure-storage-blob
azure-keyvault-secrets
azure-cosmos
pytest
requests
